Fix args passed to dcd-client and start offset of completion
This fix makes completion more finer-grained and
allows completion from the start of a symbol.

For example, before `foo.bar` would not give suggestions
for foo and would instead use omnicompletefunc. I don't think
suggestions were refined when typing each character
in `bar` either. This fix takes into account the
word that was typed, and doesn't just pass the beginning
of the match to dcd-client